requirements.txt
numpy
pandas
matplotlib
nltk
tensorflow
scikit-learn
pytorch-ignite
torcheval
optuna
plotly
streamlit
